Update:
Added a change to find_password (actual name of the non-existent "search_button" function I referenced earlier) that checks if the website is in the database or not. Returns the message "Website not in database" if the website is not in the database. Still need to evaluate if a nested try catch is needed on find_password; if not, will need to delete the nested try catch later (first one is needed). For now, I have added a placeholder on the exception to the nested try catch (print("Something wrong here...")) in case I need it later.