Update:
Deleted testing comment on find_password. I'm assuming we won't need to write if the website is already in the database, as we can just add that on the likely eventual dialog on save_password when we're trying to overwrite data on data.json. Still need to code that messagebox dialog, will probably do it later.